add basic column menu and improve sort symbol
