+++ web/src/components/ui/icons/Chevron.tsx
@@ -3,6 +3,7 @@ import {
   ChevronLeft,
   ChevronRight,
   ChevronUp,
+  ChevronsUpDown,
 } from "lucide-react";
 
 import { styled } from "@/styled-system/jsx";
@@ -11,3 +12,4 @@ export const ChevronLeftIcon = styled(ChevronLeft);
 export const ChevronRightIcon = styled(ChevronRight);
 export const ChevronUpIcon = styled(ChevronUp);
 export const ChevronDownIcon = styled(ChevronDown);
+export const ChevronUpDownIcon = styled(ChevronsUpDown);
